requirements.txt
beautifulsoup4==4.8.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
defusedxml==0.6.0
Django==2.2.6
django-allauth==0.40.0
django-crispy-forms==1.8.0
django-embed-video==1.3
gTTS==2.0.4
gTTS-token==1.1.3
idna==2.8
oauthlib==3.1.0
Pillow==6.2.1
psycopg2-binary==2.8.4
PyAudio==0.2.11
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.12.0
soupsieve==1.9.4
SpeechRecognition==3.8.1
sqlparse==0.3.0
urllib3==1.25.6
